requirements.txt
fastapi==0.68.0
uvicorn[standard]==0.15.0
python-multipart==0.0.5
pydantic==1.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
prophet==1.0.1
matplotlib==3.4.3
plotly==5.3.1
tabula-py==2.2.0
PyPDF2==2.0.0
pytest==6.2.5
httpx==0.19.0